+++ napari_accelerated_pixel_and_object_classification/_surface_vertex_classifier.py
